Further split question reproductions
